Fix to stop server errors removing all reminders
Fixed an issue where if the Android client didn't recieve a response from the server it would
remove all previously synced reminders. Also added a missing LocationLeave
enumeration constant to the server-side .Net which was causing validation
to suggest LocationLeave was not a valid action. Dropped geofence radius
to 200m.